Rename files and sycn the code
- Rename car.controller to ParkingLotController
- Rename index.controller to indexController
- Change car.js to parkingLot.js
- Rename route /cars to / parkinglot
