+++ db/cars.json
@@ -1,1 +1,1 @@
-[{"registration_no":"UT12345678","slot":"A10","park_timestamp":1679917370886},{"registration_no":"UP12345678","slot":"A8","park_timestamp":1679917417299},{"registration_no":"UK12345678","slot":"A7","park_timestamp":1679918622520},{"registration_no":"MP12345678","slot":"A4","park_timestamp":1679921557268},{"registration_no":"KA12345678","slot":"A9","park_timestamp":1679922505101}]
+[{"registration_no":"UT12345678","slot":"A10","park_timestamp":1679917370886},{"registration_no":"UP12345678","slot":"A8","park_timestamp":1679917417299},{"registration_no":"UK12345678","slot":"A7","park_timestamp":1679918622520},{"registration_no":"MP12345678","slot":"A4","park_timestamp":1679921557268}]
