+++ db/emptyslots.json
@@ -1,1 +1,1 @@
-["A1","A2","A3","A6","A5"]
+["A1","A2","A3","A6","A5","A9"]
