--- routes/cars.js
@@ -1,21 +1,0 @@
-var express = require("express");
-const CarController = require("../controller/cars.controller");
-
-const carController = new CarController();
-var router = express.Router();
-
-/* GET users listing. */
-
-router.get("/recents", carController.getRecentCars);
-
-router.get("/getslot", carController.getEmptySlot);
-
-router.get("/:registration_no", carController.getCarByRegNo);
-
-router.post("/park", carController.parkCar);
-
-router.post("/unpark", carController.unparkCar);
-
-router.get("/", carController.getAllCars);
-
-module.exports = router;
